+++ crates/matrix-sdk-ui/CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
+- Infer timeline read receipt threads for the `send_single_receipt` method from
+  the focus mode and associated `hide_threaded_events` flag. ([5325](https://github.com/matrix-org/matrix-rust-sdk/pull/5325))
 - Add `NotificationItem::room_topic` to the `NotificationItem` struct, which
   contains the topic of the room. This is useful for displaying the room topic
   in notifications. ([#5300](https://github.com/matrix-org/matrix-rust-sdk/pull/5300))